+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Add `gap` property to `FlexBox` component.
 -   Add has divider helper.
 -   Add optional `closeOnClick` property to `Select` component.
+-   Add `linkAs` prop on `SideNavigationItem`, `ListItem` and `Link` to customize the link component (can be used to inject the `Link` component from `react-router`).
 
 ### Changed